add field inspection for nested models
